+++ Utilities/ScriptingLibrary/ScriptingLibrary.cs
@@ -20,12 +20,12 @@ namespace Sheepy.PhoenixPt.ScriptingLibrary {
          Task.Run( () => Api( "eval.js", "\"Scripting Warmup\"" ) );
       }
 
-      private static readonly string[] Languages = new string[] { "js", "javascript", "ecmascript" };
+      private static readonly string[] ScriptNames = new string[] { "js", "javascript", "ecmascript" };
 
       public static object ActionMod ( string modId, Dictionary<string,object> action ) { try {
          object value = null, lang = null;
          if ( action?.TryGetValue( "eval", out value ) != true || ! ( value is string code ) ) return false;
-         if ( action?.TryGetValue( "lang", out lang ) != true || ! ( value is string l ) || ! Languages.Contains( l.Trim().ToLowerInvariant() ) ) return false;
+         if ( action?.TryGetValue( "script", out lang ) != true || ! ( value is string l ) || ! ScriptNames.Contains( l.Trim().ToLowerInvariant() ) ) return false;
          if ( Eval_Lib_Result is Exception lib_err ) return lib_err;
          Info( "Action> {0}", code );
          var result = ScriptingEngine.Eval( modId, code );